Universal Web RAG/requirements.txt
streamlit
python-dotenv

langchain
langchain-community
langchain-core
langchain-groq

sentence-transformers
faiss-cpu
